Extract captured token logic to `CaptureEmitter` for cleaner tokenization handling
