+++ src/tokenizer.js
@@ -2,6 +2,7 @@
 const reportError = require("./lib/report_error").reportError;
 const Scope = require("./scope").Scope;
 const ScopedFrameManager = require("./scoped_frame_manager").ScopedFrameManager;
+const CaptureEmitter = require("./capture_emitter").CaptureEmitter;
 var HAS_INDICES = false;
 try {
     HAS_INDICES = !!new RegExp("", "d").hasIndices;
@@ -25,6 +26,7 @@ class Tokenizer {
         this.states = rules;
         this.rootScope = new Scope(modeName || "root");
         this.scopedFrames = new ScopedFrameManager(this.rootScope);
+        this.captureEmitter = new CaptureEmitter(this);
 
         this.regExps = {};
         this.stickyRegExps = {};
@@ -107,8 +109,8 @@ class Tokenizer {
                         matchcount = 1;
                         adjustedregex = this.removeCapturingGroups(rule.regex);
                     }
-                    if (rule.tmCaptureData && HAS_INDICES)
-                        rule.tmCaptureSplitRegex = this.createSplitterRegexp(rule.regex, flag + "d");
+                    if (rule.captureData && HAS_INDICES)
+                        rule.captureSplitRegex = this.createSplitterRegexp(rule.regex, flag + "d");
                     if (!rule.splitRegex && typeof rule.token != "string")
                         splitterRurles.push(rule); // flag will be known only at the very end
                 }
@@ -251,11 +253,11 @@ class Tokenizer {
             if (lastCapture.end != null && /^\)*$/.test(src.substr(lastCapture.end)))
                 src = src.substring(0, lastCapture.start) + src.substr(lastCapture.end);
         }
-        
+
         // this is needed for regexps that can match in multiple ways
         if (src.charAt(0) != "^") src = "^" + src;
         if (src.charAt(src.length - 1) != "$") src += "$";
-        
+
         return new RegExp(src, (flag||"").replace("g", ""));
     }
 
@@ -343,24 +345,6 @@ class Tokenizer {
         return scope ? scope.attachTo(token) : token;
     }
 
-    getCaptureScopeRule(ruleScope, isScopedBegin) {
-        return {
-            ruleScope: ruleScope,
-            scope: isScopedBegin ? {type: "begin"} : null
-        };
-    }
-
-    appendRetokenizedTokens(tokens, part, retokenizeState) {
-        var retokenized = this.getLineTokens(part.value, retokenizeState || part.retokenizeState).tokens;
-        for (var j = 0; j < retokenized.length; j++) {
-            this.setTokenScope(
-                retokenized[j],
-                this.mergeRetokenizedScope(part.scope, retokenized[j].scope)
-            );
-            tokens.push(retokenized[j]);
-        }
-    }
-
     /**
      * @param {import("../ace-internal").Ace.Token[]} tokens
      * @param {import("../ace-internal").Ace.Token} token
@@ -402,170 +386,6 @@ class Tokenizer {
         return ambientScope.merge(childScope);
     }
 
-    appendStringCapturedToken(tokens, token, type, value, ruleScope, mapping, currentState, stack, merge, excludeActiveInnerScope, isTmBegin) {
-        var scope = this.getTokenScope(
-            type,
-            this.getCaptureScopeRule(ruleScope, isTmBegin),
-            mapping,
-            currentState,
-            stack,
-            excludeActiveInnerScope
-        );
-        if (merge === false) {
-            if (token.type)
-                tokens.push(token);
-            tokens.push(this.setTokenScope({
-                type: type,
-                value: value
-            }, scope));
-            return {type: null, value: ""};
-        }
-        return this.appendScopedToken(tokens, token, type, value, scope);
-    }
-
-    hasIndexedCaptures(match, type) {
-        for (var i = 1; i < match.indices.length && i <= type.length; i++) {
-            var range = match.indices[i];
-            if (range && range[0] != null && range[0] >= 0 && range[1] > range[0])
-                return true;
-        }
-        return false;
-    }
-
-    hasIndexedCaptureGaps(match, type) {
-        var fullRange = match.indices[0];
-        var position = fullRange[0];
-        for (var i = 1; i < match.indices.length && i <= type.length; i++) {
-            var range = match.indices[i];
-            if (!range || range[0] == null || range[0] < 0 || range[1] <= range[0])
-                continue;
-            if (range[0] > position)
-                return true;
-            if (range[1] > position)
-                position = range[1];
-        }
-        return fullRange[1] > position;
-    }
-
-    createCapturePart(partType, value, range) {
-        return typeof partType == "object"
-            ? Object.assign({value: value.slice(range[0], range[1])}, partType)
-            : {
-                type: partType,
-                value: value.slice(range[0], range[1])
-            };
-    }
-
-    setCapturePartScope(part, ruleScope, mapping, currentState, stack, excludeActiveInnerScope, isTmBegin) {
-        if (part.type && !part.scope)
-            this.setTokenScope(
-                part,
-                this.getTokenScope(
-                    part.type,
-                    this.getCaptureScopeRule(part.ruleScope || ruleScope, isTmBegin),
-                    mapping,
-                    currentState,
-                    stack,
-                    excludeActiveInnerScope
-                )
-            );
-    }
-
-    appendIndexedCapturedTokens(tokens, type, value, ruleScope, mapping, currentState, stack, match, excludeActiveInnerScope, isTmBegin) {
-        var fullRange = match.indices[0];
-        var position = fullRange[0];
-        var defaultType = Array.isArray(ruleScope)
-            ? ruleScope[ruleScope.length - 1]
-            : ruleScope || mapping.defaultToken || "text";
-        var defaultScope = this.getTokenScope(
-            defaultType,
-            this.getCaptureScopeRule(ruleScope, isTmBegin),
-            mapping,
-            currentState,
-            stack,
-            excludeActiveInnerScope
-        );
-        var localStack = [];
-        var currentMeta = function() {
-            return localStack[localStack.length - 1] || {
-                type: defaultType,
-                scope: defaultScope
-            };
-        };
-        var emit = function(end) {
-            if (end <= position)
-                return;
-            var meta = currentMeta();
-            tokens.push(this.setTokenScope({
-                type: meta.type || defaultType,
-                value: value.slice(position, end)
-            }, meta.scope || defaultScope));
-            position = end;
-        }.bind(this);
-
-        for (var i = 1; i < match.indices.length && i <= type.length; i++) {
-            var range = match.indices[i];
-            if (!range || range[0] == null || range[0] < 0 || range[1] <= range[0])
-                continue;
-            while (localStack.length && localStack[localStack.length - 1].end <= range[0]) {
-                emit(localStack[localStack.length - 1].end);
-                localStack.pop();
-            }
-            emit(range[0]);
-            var partType = type[i - 1];
-            if (partType == null)
-                continue;
-            var part = this.createCapturePart(partType, value, range);
-            this.setCapturePartScope(part, ruleScope, mapping, currentState, stack, excludeActiveInnerScope, isTmBegin);
-            if (part.retokenizeState) {
-                this.appendRetokenizedTokens(tokens, part);
-                position = range[1];
-                continue;
-            }
-            localStack.push({
-                end: range[1],
-                type: part.type,
-                scope: part.scope
-            });
-        }
-        while (localStack.length) {
-            emit(localStack[localStack.length - 1].end);
-            localStack.pop();
-        }
-        emit(fullRange[1]);
-        return {type: null, value: ""};
-    }
-
-    appendArrayCapturedTokens(tokens, token, type, value, ruleScope, mapping, currentState, stack, splitRegex, excludeActiveInnerScope, isTmBegin) {
-        var parts = splitRegex
-            ? this.$arrayTokens.call({splitRegex: splitRegex, tokenArray: type}, value)
-            : type.map(function(part) {
-                return Object.assign({value: value}, part);
-            });
-
-        if (typeof parts === "string") {
-            return this.appendScopedToken(
-                tokens,
-                token,
-                parts,
-                value,
-                this.getTokenScope(parts, this.getCaptureScopeRule(ruleScope, isTmBegin), mapping, currentState, stack, excludeActiveInnerScope)
-            );
-        }
-
-        if (token.type)
-            tokens.push(token);
-        for (var i = 0; i < parts.length; i++) {
-            this.setCapturePartScope(parts[i], ruleScope, mapping, currentState, stack, excludeActiveInnerScope, isTmBegin);
-            tokens.push(parts[i]);
-            if (parts[i].retokenizeState) {
-                tokens.pop();
-                this.appendRetokenizedTokens(tokens, parts[i]);
-            }
-        }
-        return {type: null, value: ""};
-    }
-
     /**
      * @param {import("../ace-internal").Ace.Token[]} tokens
      * @param {import("../ace-internal").Ace.Token} token
@@ -578,54 +398,7 @@ class Tokenizer {
      * @param {RegExp} splitRegex
      */
     appendCapturedTokens(tokens, token, type, value, ruleScope, mapping, currentState, stack, splitRegex, merge, excludeActiveInnerScope, emitGaps, isTmBegin) {
-        if (!value || !type)
-            return token;
-
-        if (typeof type === "string")
-            return this.appendStringCapturedToken(
-                tokens,
-                token,
-                type,
-                value,
-                ruleScope,
-                mapping,
-                currentState,
-                stack,
-                merge,
-                excludeActiveInnerScope,
-                isTmBegin
-            );
-
-        if (!Array.isArray(type))
-            type = [type];
-
-        if (emitGaps && splitRegex && HAS_INDICES) {
-            var match = splitRegex.exec(value);
-            if (match && match.indices && match.indices[0]) {
-                if (token.type)
-                    tokens.push(token);
-                token = {type: null, value: ""};
-
-                if (
-                    this.hasIndexedCaptures(match, type) &&
-                    this.hasIndexedCaptureGaps(match, type)
-                )
-                    return this.appendIndexedCapturedTokens(
-                        tokens,
-                        type,
-                        value,
-                        ruleScope,
-                        mapping,
-                        currentState,
-                        stack,
-                        match,
-                        excludeActiveInnerScope,
-                        isTmBegin
-                    );
-            }
-        }
-
-        return this.appendArrayCapturedTokens(
+        return this.captureEmitter.append(
             tokens,
             token,
             type,
@@ -635,8 +408,11 @@ class Tokenizer {
             currentState,
             stack,
             splitRegex,
+            merge,
             excludeActiveInnerScope,
-            isTmBegin
+            emitGaps,
+            isTmBegin,
+            HAS_INDICES
         );
     }
 
@@ -893,7 +669,7 @@ class Tokenizer {
                     } else {
                         currentState = rule.next(currentState, stack);
                     }
-                    
+
                     state = this.states[currentState];
                     if (!state) {
                         this.reportError("state doesn't exist", currentState);
@@ -1099,7 +875,7 @@ class Tokenizer {
                 break;
             currentState = this.scopedFrames.popSpecific(currentState, stack, finalFrame);
         }
-        
+
         if (stack.length > 1) {
             if (stack[0] !== currentState)
                 stack.unshift("#tmp", currentState);
